Add integration tests for cases routes
- GET cases/
- GET cases/:id
- GET and POST cases/:id/messages
- GET and POST cases/:id/events
- Add config option for sequelize logging
