+++ .gitignore
@@ -1,3 +1,4 @@
+.env-test
 .env
 api/node_modules/
 api/dist/
